Overall scripts edits to meetup espressif master branch.

+++ build.sh
@@ -12,31 +12,44 @@ fi
 
 export TARGET="all"
 BUILD_TYPE="all"
+BUILD_DEBUG="default"
 SKIP_ENV=0
 COPY_OUT=0
 ARCHIVE_OUT=1
-DEPLOY_OUT=0
+if [ -z $DEPLOY_OUT ]; then
+    DEPLOY_OUT=0
+fi
 
 function print_help() {
-    echo "Usage: build.sh [-s] [-A <arduino_branch>] [-I <idf_branch>] [-i <idf_commit>] [-c <path>] [-t <target>] [-b <build|menuconfig|reconfigure|idf-libs|copy-bootloader|mem-variant>] [config ...]"
+    echo "Usage: build.sh [-s] [-A <arduino_branch>] [-I <idf_branch>] [-D <debug_level>] [-i <idf_commit>] [-c <path>] [-t <target>] [-b <build|menuconfig|reconfigure|idf-libs|copy-bootloader|mem-variant>] [config ...]"
     echo "       -s     Skip installing/updating of ESP-IDF and all components"
     echo "       -A     Set which branch of arduino-esp32 to be used for compilation"
     echo "       -I     Set which branch of ESP-IDF to be used for compilation"
     echo "       -i     Set which commit of ESP-IDF to be used for compilation"
     echo "       -e     Archive the build to dist"
+    echo "       -d     Deploy the build to github arduino-esp32"
+    echo "       -D     Debug level to be set to ESP-IDF. One of default,none,error,warning,info,debug or verbose"
+    echo "       -c     Set the arduino-esp32 folder to copy the result to. ex. '$HOME/Arduino/hardware/espressif/esp32'"
     echo "       -t     Set the build target(chip) ex. 'esp32s3' or select multiple targets(chips) by separating them with comma ex. 'esp32,esp32s3,esp32c3'"
     echo "       -b     Set the build type. ex. 'build' to build the project and prepare for uploading to a board"
     echo "       ...    Specify additional configs to be applied. ex. 'qio 80m' to compile for QIO Flash@80MHz. Requires -b"
     exit 1
 }
 
-while getopts ":A:I:i:c:t:b:sde" opt; do
+while getopts ":A:I:i:c:t:b:D:sde" opt; do
     case ${opt} in
         s )
             SKIP_ENV=1
             ;;
+        d )
+            DEPLOY_OUT=1
+            ;;
         e )
             ARCHIVE_OUT=1
+            ;;
+        c )
+            export ESP32_ARDUINO="$OPTARG"
+            COPY_OUT=1
             ;;
         A )
             export AR_BRANCH="$OPTARG"
@@ -47,16 +60,19 @@ while getopts ":A:I:i:c:t:b:sde" opt; do
         i )
             export IDF_COMMIT="$OPTARG"
             ;;
+        D )
+            BUILD_DEBUG="$OPTARG"
+            ;;
         t )
             IFS=',' read -ra TARGET <<< "$OPTARG"
             ;;
         b )
             b=$OPTARG
-            if [ "$b" != "build" ] &&
-               [ "$b" != "menuconfig" ] &&
-               [ "$b" != "reconfigure" ] &&
-               [ "$b" != "idf-libs" ] &&
-               [ "$b" != "copy-bootloader" ] &&
+            if [ "$b" != "build" ] && 
+               [ "$b" != "menuconfig" ] && 
+               [ "$b" != "reconfigure" ] && 
+               [ "$b" != "idf-libs" ] && 
+               [ "$b" != "copy-bootloader" ] && 
                [ "$b" != "mem-variant" ]; then
                 print_help
             fi
@@ -100,6 +116,10 @@ else
     source ./tools/config.sh
 fi
 
+if [ -f "$AR_MANAGED_COMPS/espressif__esp-sr/.component_hash" ]; then
+    rm -rf $AR_MANAGED_COMPS/espressif__esp-sr/.component_hash
+fi
+
 if [ "$BUILD_TYPE" != "all" ]; then
     if [ "$TARGET" = "all" ]; then
         echo "ERROR: You need to specify target for non-default builds"
@@ -123,8 +143,8 @@ if [ "$BUILD_TYPE" != "all" ]; then
             # Skip building for targets that are not in the $TARGET array
             continue
         fi
-
-        configs="configs/defconfig.common;configs/defconfig.$target"
+                
+        configs="configs/defconfig.common;configs/defconfig.$target;configs/defconfig.debug_$BUILD_DEBUG"
         for defconf in `echo "$target_json" | jq -c '.features[]' | tr -d '"'`; do
             configs="$configs;configs/defconfig.$defconf"
         done
@@ -179,7 +199,7 @@ for target_json in `jq -c '.targets[]' configs/builds.json`; do
             continue
         fi
     fi
-
+    
     # Skip chips that should not be a part of the final libs
     # WARNING!!! this logic needs to be updated when cron builds are split into jobs
     if [ "$TARGET" = "all" ] && [ $target_skip -eq 1 ]; then
@@ -190,7 +210,7 @@ for target_json in `jq -c '.targets[]' configs/builds.json`; do
     echo "* Target: $target"
 
     # Build Main Configs List
-    main_configs="configs/defconfig.common;configs/defconfig.$target"
+    main_configs="configs/defconfig.common;configs/defconfig.$target;configs/defconfig.debug_$BUILD_DEBUG"
     for defconf in `echo "$target_json" | jq -c '.features[]' | tr -d '"'`; do
         main_configs="$main_configs;configs/defconfig.$defconf"
     done
@@ -201,17 +221,36 @@ for target_json in `jq -c '.targets[]' configs/builds.json`; do
         idf_libs_configs="$idf_libs_configs;configs/defconfig.$defconf"
     done
 
+    if [ -f "$AR_MANAGED_COMPS/espressif__esp-sr/.component_hash" ]; then
+        rm -rf $AR_MANAGED_COMPS/espressif__esp-sr/.component_hash
+    fi
     echo "* Build IDF-Libs: $idf_libs_configs"
     rm -rf build sdkconfig
     idf.py -DIDF_TARGET="$target" -DSDKCONFIG_DEFAULTS="$idf_libs_configs" idf-libs
     if [ $? -ne 0 ]; then exit 1; fi
 
+    if [ "$target" == "esp32s3" ]; then
+        idf.py -DIDF_TARGET="$target" -DSDKCONFIG_DEFAULTS="$idf_libs_configs" srmodels_bin
+        if [ $? -ne 0 ]; then exit 1; fi
+        AR_SDK="$AR_TOOLS/esp32-arduino-libs/$target"
+        # sr model.bin
+        if [ -f "build/srmodels/srmodels.bin" ]; then
+            echo "$AR_SDK/esp_sr"
+            mkdir -p "$AR_SDK/esp_sr"
+            cp -f "build/srmodels/srmodels.bin" "$AR_SDK/esp_sr/"
+            cp -f "partitions.csv" "$AR_SDK/esp_sr/"
+        fi
+    fi
     # Build Bootloaders
     for boot_conf in `echo "$target_json" | jq -c '.bootloaders[]'`; do
         bootloader_configs="$main_configs"
         for defconf in `echo "$boot_conf" | jq -c '.[]' | tr -d '"'`; do
             bootloader_configs="$bootloader_configs;configs/defconfig.$defconf";
         done
+
+        if [ -f "$AR_MANAGED_COMPS/espressif__esp-sr/.component_hash" ]; then
+            rm -rf $AR_MANAGED_COMPS/espressif__esp-sr/.component_hash
+        fi
 
         echo "* Build BootLoader: $bootloader_configs"
         rm -rf build sdkconfig
@@ -226,6 +265,9 @@ for target_json in `jq -c '.targets[]' configs/builds.json`; do
             mem_configs="$mem_configs;configs/defconfig.$defconf";
         done
 
+        if [ -f "$AR_MANAGED_COMPS/espressif__esp-sr/.component_hash" ]; then
+            rm -rf $AR_MANAGED_COMPS/espressif__esp-sr/.component_hash
+        fi
         echo "* Build Memory Variant: $mem_configs"
         rm -rf build sdkconfig
         idf.py -DIDF_TARGET="$target" -DSDKCONFIG_DEFAULTS="$mem_configs" mem-variant
@@ -264,23 +306,33 @@ for component in `ls "$AR_MANAGED_COMPS"`; do
     fi
 done
 
+# update package_esp32_index.template.json
+if [ "$BUILD_TYPE" = "all" ]; then
+    python3 ./tools/gen_tools_json.py -i "$IDF_PATH" -j "$AR_COMPS/arduino/package/package_esp32_index.template.json" -o "$AR_OUT/"
+    python3 ./tools/gen_tools_json.py -i "$IDF_PATH" -o "$TOOLS_JSON_OUT/"
+    if [ $? -ne 0 ]; then exit 1; fi
+fi
 export IDF_COMMIT=$(git -C "$IDF_PATH" rev-parse --short HEAD)
 
-# Generate PlatformIO library manifest file
+# Generate PlatformIO manifest file
 if [ "$BUILD_TYPE" = "all" ]; then
-    python3 ./tools/gen_pio_lib_manifest.py -o "$TOOLS_JSON_OUT/" -s "v$IDF_VERSION" -c "$IDF_COMMIT"
+    pushd $IDF_PATH
+    ibr=$(git describe --all --exact-match 2>/dev/null)
+    ic=$(git -C "$IDF_PATH" rev-parse --short HEAD)
+    popd
+    python3 ./tools/gen_platformio_manifest.py -o "$TOOLS_JSON_OUT/" -s "$ibr" -c "$ic"
     if [ $? -ne 0 ]; then exit 1; fi
 fi
 
+# copy everything to arduino-esp32 installation
+if [ $COPY_OUT -eq 1 ] && [ -d "$ESP32_ARDUINO" ]; then
+    ./tools/copy-to-arduino.sh
+    if [ $? -ne 0 ]; then exit 1; fi
+fi
 AR_VERSION=$(jq -c '.version' "$AR_COMPS/arduino/package.json" | tr -d '"')
 AR_VERSION_UNDERSCORE=`echo "$AR_VERSION" | tr . _`
 
-# Generate PlatformIO framework manifest file
-rm -rf "$AR_ROOT/package.json"
-if [ "$BUILD_TYPE" = "all" ]; then
-    python3 ./tools/gen_pio_frmwk_manifest.py -o "$AR_ROOT/" -s "v$AR_VERSION" -c "$IDF_COMMIT"
-    if [ $? -ne 0 ]; then exit 1; fi
-fi
+
 
 # Generate core_version.h
 rm -rf "$AR_ROOT/core_version.h"
@@ -289,6 +341,11 @@ echo "#define ARDUINO_ESP32_GIT_VER 0x$AR_Commit_short
 #define ARDUINO_ESP32_RELEASE_$AR_VERSION_UNDERSCORE
 #define ARDUINO_ESP32_RELEASE \"$AR_VERSION_UNDERSCORE\"" >> "$AR_ROOT/core_version.h"
 
+# push changes to esp32-arduino-libs and create pull request into arduino-esp32
+if [ $DEPLOY_OUT -eq 1 ]; then
+    ./tools/push-to-arduino.sh
+    if [ $? -ne 0 ]; then exit 1; fi
+fi
 # archive the build
 if [ $ARCHIVE_OUT -eq 1 ]; then
     ./tools/archive-build.sh "$TARGET"
